llm-flask/requirements.txt
Flask
pdf2image
Pillow
google-cloud-vision
faiss-cpu
sentence-transformers
numpy
google-genai
google-cloud-storage
langchain
langchain-community